Remove the get_phrase_cfws_only_raises test
This test's name does not start with test_, which prevents it from
running. Apparently it was accidentally added this way.

None of the callers of get_phrase() assume it can raise exceptions, so
simply dropping the test seems best. The fact that the phrase is invalid
still gets recorded in defects.