bpo-39100: _header_value_parser: do not treat a Group as invalid-mailbox
When an address in an address-list has garbage at the end, its
child mailbox is changed into an invalid-mailbox by overriding its
token_type. However, that does not take into account that an address may
also contain a Group instead of a single mailbox. In that case,
overwriting token_type leads to undesirable results, e.g. parsing an
email with the following 'To' header:

unlisted-recipients:; (no To-header on input)

raises an AttributeError from trying to treat the Group as a Mailbox.

Fix this by making the invalid-mailbox replacement only for an actual
mailbox. For a group, keep it as is and just append the trailing garbage
to the end.